# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Branch-Out API
  version: 0.5.0
  description: PostgreSQL-backed discovery, GitHub authentication, member profiles, owner-managed opening lifecycles, and proof-led application submission for Branch-Out.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/openings:
    get:
      operationId: listOpenings
      summary: List project openings
      parameters:
        - name: query
          in: query
          description: Case-insensitive terms that must all match searchable opening fields.
          schema:
            type: string
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/Role'
        - name: compensation
          in: query
          schema:
            $ref: '#/components/schemas/Compensation'
        - name: commitment
          in: query
          schema:
            $ref: '#/components/schemas/CommitmentBand'
      responses:
        '200':
          description: Matching project openings.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Opening'
                  meta:
                    type: object
                    required: [count]
                    properties:
                      count:
                        type: integer
                        minimum: 0
        '400':
          description: A structured filter is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: The catalogue could not be loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      operationId: createOpeningDraft
      summary: Create an opening draft owned by the authenticated member
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpeningDraftInput'
      responses:
        '201':
          description: The private owner draft was created.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/ManagedOpening'
        '400':
          description: The request or an opening field is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: No valid session is present.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: The authenticated member must complete a profile first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: The draft could not be created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/openings/mine:
    get:
      operationId: listOwnedOpenings
      summary: List openings owned by the authenticated member
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The member's openings, including private drafts.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ManagedOpening'
                  meta:
                    type: object
                    required: [count]
                    properties:
                      count:
                        type: integer
                        minimum: 0
        '401':
          description: No valid session is present.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: The owned openings could not be loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/openings/{id}:
    put:
      operationId: updateOpeningDraft
      summary: Replace an opening draft owned by the authenticated member
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpeningDraftInput'
      responses:
        '200':
          description: The private owner draft was updated.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/ManagedOpening'
        '400':
          description: The request or an opening field is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: No valid session is present.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: No editable draft with this ID belongs to the member.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: The authenticated member must complete a profile first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: The draft could not be updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/openings/{id}/publish:
    post:
      operationId: publishOwnedOpening
      summary: Publish an opening draft owned by the authenticated member
      description: Atomically changes an owned draft to published so it becomes publicly discoverable.
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The opening is published.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/ManagedOpening'
        '401':
          description: No valid session is present.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: No owned draft is available for this transition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: The opening could not be published.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/openings/{id}/close:
    post:
      operationId: closeOwnedOpening
      summary: Close a published opening owned by the authenticated member
      description: Atomically changes an owned published opening to closed so it leaves public discovery.
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The opening is closed.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/ManagedOpening'
        '401':
          description: No valid session is present.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: No owned published opening is available for this transition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: The opening could not be closed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/openings/{id}/application:
    get:
      operationId: getOwnApplication
      summary: Get the authenticated member's application for an opening
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The member's private or submitted application.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Application'
        '401':
          description: No valid session is present.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: The member has no application for this opening.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: The application could not be loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      operationId: saveOwnApplicationDraft
      summary: Create or replace the authenticated member's private application draft
      description: Requires a completed profile and a published opening owned by someone else. Submitted applications are immutable.
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationInput'
      responses:
        '200':
          description: The private application draft was saved.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Application'
        '400':
          description: The request or an application field is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: No valid session is present.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: A profile is required, the opening is unavailable or owned by the applicant, or the application is already submitted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: The application draft could not be saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/openings/{id}/application/submit:
    post:
      operationId: submitOwnApplication
      summary: Submit the authenticated member's private application
      description: Atomically changes a private draft to submitted. Submission is irreversible in this milestone.
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The application was submitted and is now immutable.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Application'
        '401':
          description: No valid session is present.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: The application or opening is not available for submission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: The application could not be submitted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/auth/github/start:
    get:
      operationId: startGitHubAuthentication
      summary: Start GitHub OAuth authentication
      description: Creates a single-use OAuth attempt with PKCE and redirects to GitHub.
      responses:
        '303':
          description: Continue authentication at GitHub.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '500':
          description: Authentication could not be started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: GitHub OAuth credentials are not configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/auth/github/callback:
    get:
      operationId: finishGitHubAuthentication
      summary: Complete GitHub OAuth authentication
      description: Validates browser-bound state and PKCE, creates a session, and redirects to the frontend.
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
      responses:
        '303':
          description: Authentication completed, was denied, or failed; inspect the frontend auth query value.
          headers:
            Location:
              schema:
                type: string
                format: uri
        '500':
          description: The configured frontend redirect URL is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/session:
    get:
      operationId: getCurrentSession
      summary: Get the authenticated Branch-Out user
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The current authenticated user.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: No valid session is present.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: The session could not be loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      operationId: deleteCurrentSession
      summary: Log out the current user
      security:
        - sessionCookie: []
      responses:
        '204':
          description: The session was deleted or no session was present.
        '500':
          description: The session could not be deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/profile:
    get:
      operationId: getCurrentProfile
      summary: Get the authenticated member's collaboration profile
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The current member's profile.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Profile'
        '401':
          description: No valid session is present.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: The authenticated member has not completed a profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: The profile could not be loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      operationId: putCurrentProfile
      summary: Create or replace the authenticated member's collaboration profile
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileInput'
      responses:
        '200':
          description: The saved profile.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Profile'
        '400':
          description: The request or a profile field is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: No valid session is present.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: The profile could not be saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: branch_out_session
  schemas:
...
    Role:
      type: string
      enum: [Engineering, Design, Research]
    Compensation:
      type: string
      enum: [Paid, Fixed bounty, Revenue share, Portfolio]
    CommitmentBand:
      type: string
      enum: [Under 6 hrs/week, 6–8 hrs/week, 8+ hrs/week]
    Opening:
      type: object
      required:
        - id
        - title
        - summary
        - skills
        - role
        - compensation
        - commitment
        - commitmentBand
        - duration
        - timezone
        - freshness
        - stage
        - desiredOutcome
        - firstMilestone
        - ownerContribution
        - ownerName
        - ownerSignal
        - confidentiality
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        skills:
          type: array
          items:
            type: string
        role:
          $ref: '#/components/schemas/Role'
        compensation:
          $ref: '#/components/schemas/Compensation'
        commitment:
          type: string
        commitmentBand:
          $ref: '#/components/schemas/CommitmentBand'
        duration:
          type: string
        timezone:
          type: string
        freshness:
          type: string
        stage:
          type: string
        desiredOutcome:
          type: string
        firstMilestone:
          type: string
        ownerContribution:
          type: string
        ownerName:
          type: string
        ownerSignal:
          type: string
        confidentiality:
          type: string
    ManagedOpening:
      allOf:
        - $ref: '#/components/schemas/Opening'
        - type: object
          required: [publicationStatus]
          properties:
            publicationStatus:
              type: string
              enum: [draft, published, closed]
    OpeningDraftInput:
      type: object
      additionalProperties: false
      required:
        - projectName
        - problem
        - role
        - skills
        - commitment
        - duration
        - timezone
        - compensation
        - firstMilestone
        - ownerContribution
        - confidentiality
      properties:
        projectName:
          type: string
          minLength: 3
          maxLength: 80
        problem:
          type: string
          minLength: 20
          maxLength: 240
        role:
          type: string
          enum: [Frontend engineer, Backend engineer, Product designer, UX researcher]
        skills:
          type: array
          minItems: 1
          maxItems: 12
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 40
        commitment:
          type: string
          enum: [Under 6 hrs/week, 6–8 hrs/week, 8+ hrs/week]
        duration:
          type: string
          enum: [2–4 weeks, 5–8 weeks, 2–3 months]
        timezone:
          type: string
          minLength: 3
          maxLength: 80
        compensation:
          type: string
          enum: [Paid, Fixed bounty, Revenue share, Unpaid / portfolio]
        firstMilestone:
          type: string
          minLength: 20
          maxLength: 500
        ownerContribution:
          type: string
          minLength: 20
          maxLength: 500
        confidentiality:
          type: string
          enum: [Public, Limited details, Confidential after agreement]
    ApplicationInput:
      type: object
      additionalProperties: false
      required:
        - message
        - workSampleUrl
        - workSampleContext
        - availability
        - availabilityConfirmed
        - proposedContribution
      properties:
        message:
          type: string
          minLength: 30
          maxLength: 1000
        workSampleUrl:
          type: string
          format: uri
          maxLength: 2048
        workSampleContext:
          type: string
          minLength: 20
          maxLength: 500
        availability:
          type: string
          minLength: 3
          maxLength: 160
        availabilityConfirmed:
          type: boolean
          const: true
        proposedContribution:
          type: string
          minLength: 20
          maxLength: 500
    Application:
      type: object
      required: [id, openingId, input, status, submittedAt, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        openingId:
          type: string
        input:
          $ref: '#/components/schemas/ApplicationInput'
        status:
          type: string
          enum: [draft, submitted]
        submittedAt:
          type:
            - string
            - 'null'
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            field:
              type: string
    User:
      type: object
      required:
        - id
        - githubUserId
        - githubLogin
        - avatarUrl
        - profileUrl
      properties:
        id:
          type: integer
          format: int64
        githubUserId:
          type: integer
          format: int64
        githubLogin:
          type: string
          minLength: 1
          maxLength: 39
        displayName:
          type:
            - string
            - 'null'
          maxLength: 100
        avatarUrl:
          type: string
          format: uri
        profileUrl:
          type: string
          format: uri
...
    Profile:
      allOf:
        - $ref: '#/components/schemas/ProfileInput'
        - type: object
          required: [userId, githubUrl, createdAt, updatedAt]
          properties:
            userId:
              type: integer
              format: int64
            githubUrl:
              type: string
              format: uri
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time